oth: update test config
