+++ .env.test
@@ -13,4 +13,4 @@ WEBHOOK_DOMAIN="example.com"
 # Telegram bot webhook path if it's necessary to use with webhooks (optional)
 WEBHOOK_PATH="/chat-adviser-bot"
 # Telegram bot webhook port if it's necessary to use with webhooks (optional)
-WEBHOOK_PORT=5434
+WEBHOOK_PORT=3001
