+++ .github/workflows/main.yml
@@ -52,7 +52,7 @@ jobs:
       - name: Merge test coverage
         run: npm run merge:cov
       - name: Check test coverage badges
-        run: npm run make-badges -- --ci
+        run: npm run make:badges -- --ci
 
   deploy:
     environment: production
