+++ package.json
@@ -6,30 +6,28 @@
   "license": "MIT",
   "scripts": {
     "build": "eslint \"{src,apps,libs,test}/**/*.ts\" && nest build",
+    "dev:env:up": "docker compose -f compose.dev.yml up --force-recreate --remove-orphans --wait -d && prisma migrate dev",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "make:badges": "istanbul-badges-readme --coverageDir=\"./coverage-total\" --silent",
+    "merge:cov": "nyc report --report-dir coverage-total --reporter=json-summary --reporter=html --reporter=text-summary",
     "start": "npm run dev:env:up && nest start",
     "start:debug": "npm run dev:env:up && nest start --debug --watch",
     "start:dev": "npm run dev:env:up && nest start --watch",
     "start:prod": "node dist/main",
     "start:prod:docker": "docker compose up --build --force-recreate --remove-orphans -d",
-    "test": "npm run test:env:up && jest && npm run test:env:down",
-    "test:cov": "npm run test:env:up && jest --coverage && npm run test:env:down",
-    "test:watch": "npm run test:env:up && jest --watch && npm run test:env:down",
-    "test:e2e": "npm run test:env:up && jest --config jest-e2e.config.ts && npm run test:env:down",
-    "test:e2e:cov": "npm run test:env:up && jest --config jest-e2e.config.ts --coverage && npm run test:env:down",
-    "test:e2e:watch": "npm run test:env:up && jest --config jest-e2e.config.ts --watch && npm run test:env:down",
-    "test:total:cov": "npm run test:cov && npm run test:e2e:cov && npm run merge:cov && npm run make-badges",
-    "dev:env:up": "docker compose -f compose.dev.yml up --force-recreate --remove-orphans --wait -d && prisma migrate dev",
-    "make-badges": "istanbul-badges-readme --coverageDir=\"./coverage-total\" --silent",
-    "merge:cov": "nyc report --report-dir coverage-total --reporter=json-summary --reporter=html --reporter=text-summary",
-    "test:env:down": "docker compose --env-file=.env.test -f compose.test.yml down",
-    "test:env:up": "docker compose --env-file=.env.test -f compose.test.yml up --force-recreate --remove-orphans --wait -d && dotenv -e .env.test -- prisma migrate deploy"
+    "test": "jest",
+    "test:cov": "jest --coverage",
+    "test:watch": "jest --watch",
+    "test:e2e": "jest --config jest-e2e.config.ts",
+    "test:e2e:cov": "jest --config jest-e2e.config.ts --coverage",
+    "test:e2e:watch": "jest --config jest-e2e.config.ts --watch",
+    "test:total:cov": "npm run test:cov && npm run test:e2e:cov && npm run merge:cov && npm run make:badges"
   },
   "dependencies": {
-    "@nestjs/common": "10.3.1",
-    "@nestjs/config": "3.1.1",
-    "@nestjs/core": "10.3.1",
-    "@nestjs/schedule": "4.0.0",
+    "@nestjs/common": "10.3.2",
+    "@nestjs/config": "3.2.0",
+    "@nestjs/core": "10.3.2",
+    "@nestjs/schedule": "4.0.1",
     "@prisma/client": "5.9.1",
     "date-fns": "2.30.0",
     "date-fns-tz": "2.0.0",
@@ -41,21 +39,20 @@
     "telegraf": "4.15.3"
   },
   "devDependencies": {
-    "@nestjs/cli": "10.3.1",
-    "@nestjs/schematics": "10.1.0",
-    "@nestjs/testing": "10.3.1",
+    "@nestjs/cli": "10.3.2",
+    "@nestjs/schematics": "10.1.1",
+    "@nestjs/testing": "10.3.2",
     "@types/jest": "29.5.12",
     "@types/lodash.escaperegexp": "4.1.9",
     "@types/node": "20.11.16",
     "@types/supertest": "6.0.2",
     "@typescript-eslint/eslint-plugin": "6.21.0",
     "@typescript-eslint/parser": "6.21.0",
-    "dotenv-cli": "7.3.0",
     "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-import": "2.29.1",
     "eslint-plugin-jest": "27.6.3",
-    "eslint-plugin-jsdoc": "48.0.5",
+    "eslint-plugin-jsdoc": "48.0.6",
     "eslint-plugin-optimize-regex": "1.2.1",
     "eslint-plugin-prettier": "5.1.3",
     "eslint-plugin-promise": "6.1.1",
