+++ src/language/translations/ru.json
@@ -22,7 +22,7 @@
     "commandForAdmins": "Эта команда доступна только администраторам.",
     "commandForPrivateChat": "Напишите эту команду мне в <a href=\"{{BOT_LINK}}\">личные сообщения</a>, и я помогу настроить Ваши чаты.",
     "commandNotForPrivateChats": "Эта команда не для личных чатов.",
-    "help": "Здравствуйте! Я бот, который помогает модерировать чаты.\n\n<b>С чего начать:</b>\n1. добавьте меня в чат\n2. предоставьте мне права администратора\n3. напишите в <a href=\"{{BOT_LINK}}\">личные сообщения</a> команду /mychats, и я помогу настроить Ваш чат\n\n<b>Список функций:</b> голосование за блокировку, фильтр матов, запрет добавления ботов, поддержка различных языков.\nВ момент настройки я расскажу о каждой функции более подробно.\n\nВ любой момент можно вызвать это сообщение еще раз при помощи команды /help.",
+    "help": "Здравствуйте! Я бот, который помогает модерировать чаты.\n\n<b>С чего начать:</b>\n1. добавьте меня в чат\n2. предоставьте мне права администратора\n3. напишите в <a href=\"{{BOT_LINK}}\">личные сообщения</a> команду /mychats, и я помогу настроить Ваш чат\n\n<b>Список функций:</b> голосование за блокировку, фильтр матов, запрет добавления ботов, поддержка различных языков и т.д.\nВ момент настройки я расскажу о каждой функции более подробно.\n\nВ любой момент можно вызвать это сообщение еще раз при помощи команды /help.",
     "needAdminPermissions": "Для работы этой функции мне необходимы права администратора."
   },
   "language": {
