+++ test/channel-message-filter.e2e-spec.ts
@@ -3,11 +3,16 @@ import { Test } from "@nestjs/testing";
 import { http, HttpResponse } from "msw";
 import request from "supertest";
 import type { App } from "supertest/types";
-import { server } from "test/utils/setup";
+import { server } from "test/utils/setup-after-env";
 
 import { AppModule } from "../src/app.module";
-import { ASYNC_REQUEST_DELAY, TELEGRAM_BOT_API_BASE_URL, TEST_WEBHOOK_BASE_URL, TEST_WEBHOOK_PATH } from "./constants";
 import * as fixtures from "./fixtures/channel-message-filter";
+import {
+  ASYNC_REQUEST_DELAY,
+  TELEGRAM_API_BASE_URL,
+  TEST_WEBHOOK_BASE_URL,
+  TEST_WEBHOOK_PATH,
+} from "./utils/constants";
 import { createDbSupergroupChat } from "./utils/database";
 import { sleep } from "./utils/sleep";
 
@@ -25,7 +30,7 @@ describe("ChannelMessageFilterModule (e2e)", () => {
   it("should filter channel messages in a new supergroup chat", async () => {
     let banChatSenderChatPayload;
     server.use(
-      http.post(`${TELEGRAM_BOT_API_BASE_URL}/banChatSenderChat`, async (info) => {
+      http.post(`${TELEGRAM_API_BASE_URL}/banChatSenderChat`, async (info) => {
         banChatSenderChatPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
@@ -54,7 +59,7 @@ describe("ChannelMessageFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let banChatSenderChatPayload;
     server.use(
-      http.post(`${TELEGRAM_BOT_API_BASE_URL}/banChatSenderChat`, async (info) => {
+      http.post(`${TELEGRAM_API_BASE_URL}/banChatSenderChat`, async (info) => {
         banChatSenderChatPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
@@ -71,7 +76,7 @@ describe("ChannelMessageFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let editMessageTextPayload;
     server.use(
-      http.post(`${TELEGRAM_BOT_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
@@ -88,7 +93,7 @@ describe("ChannelMessageFilterModule (e2e)", () => {
     await createDbSupergroupChat();
     let editMessageTextPayload;
     server.use(
-      http.post(`${TELEGRAM_BOT_API_BASE_URL}/editMessageText`, async (info) => {
+      http.post(`${TELEGRAM_API_BASE_URL}/editMessageText`, async (info) => {
         editMessageTextPayload = await info.request.json();
         return HttpResponse.json({ ok: true });
       }),
